backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
mysql-connector-python==8.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic[email]==2.6.4
httpx==0.27.0

# ── Web Scraping + ETL module ──────────────────────────────────────────────
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.1
